pyproject.toml
[project]
name = "agent-smith"
version = "0.1.0"
description = "Lightweight development environment for Agent Skills"
authors = [
    {name = "Sean Betts", email = "seanbetts@icloud.com"}
]
requires-python = ">=3.11"

# Runtime dependencies for skills with bundled scripts
dependencies = [
    # PDF processing (pdf skill)
    "pypdf>=4.0.0",
    # PowerPoint processing (pptx skill)
    "python-pptx>=0.6.23",
    # Image manipulation (used by pptx skill)
    "Pillow>=10.0.0",
    # Safe XML parsing (used by docx/pptx ooxml scripts)
    "defusedxml>=0.7.1",
    # HTTP library (for custom scripts)
    "requests>=2.31.0",
    # YAML processing (folder-config skill)
    "PyYAML>=6.0",
]

[dependency-groups]
dev = [
    # Agent Skills validation tool (dev only)
    "skills-ref @ git+https://github.com/agentskills/agentskills.git#subdirectory=skills-ref",
    # Testing infrastructure
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.5.0",  # Parallel test execution
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",                    # Verbose output
    "--strict-markers",      # Raise error on unknown markers
    "--tb=short",           # Shorter traceback format
    "--cov=scripts",        # Coverage for scripts/
    "--cov=skills",         # Coverage for skills/*/scripts/
    "--cov-report=term-missing",  # Show missing lines
    "--cov-report=html",    # HTML coverage report
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
